Never autoClose file descriptors
Closing the file descriptors is racy. If a Readable is not fully
consumed then the `autoClose` will never hit and our manual
attempt will be ok. Most of the cases the automatic closing is
fine, even if we manually close the same fd. However, sometimes
it leads to a segfault. On the other hand, if we do not close the
files at all then then GC will complain and we might even run out
of file descriptors before the GC runs.

```
FATAL ERROR: v8::HandleScope::CreateHandle() Cannot create a handle without a HandleScope
 1: 0x9ef190 node::Abort() [node]
 2: 0x9f13b2 node::OnFatalError(char const*, char const*) [node]
 3: 0xb5d86a v8::Utils::ReportApiFailure(char const*, char const*) [node]
 4: 0xcd2a6a v8::internal::HandleScope::Extend(v8::internal::Isolate*) [node]
 5: 0xce8d0c v8::internal::Factory::NewRawOneByteString(int, v8::internal::AllocationType) [node]
 6: 0xce8d51 v8::internal::Factory::NewStringFromOneByte(v8::internal::Vector<unsigned char const> const&, v8::internal::AllocationType) [node]
 7: 0xb75faf v8::String::NewFromOneByte(v8::Isolate*, unsigned char const*, v8::NewStringType, int) [node]
 8: 0x962fb5 node::UVException(v8::Isolate*, int, char const*, char const*, char const*, char const*) [node]
 9: 0x9f30d1  [node]
10: 0x13093f5  [node]
11: 0x130d8e1  [node]
12: 0x1320188  [node]
13: 0x130e26b uv_run [node]
14: 0xa31ec3 node::NodeMainInstance::Run() [node]
15: 0x9c1cc8 node::Start(int, char**) [node]
16: 0x7f39557e71a3 __libc_start_main [/lib64/libc.so.6]
17: 0x95c085  [node]
```